+++ _boards/tqma8mpxl/examples/demo_apps/hello_world/app.h
@@ -1,0 +1,23 @@
+//*****************************************************************************
+/*!
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
+ * Copyright 2020 NXP
+ * All rights reserved.
+ *
+ * Copyright (c) 2021-2026 TQ-Systems GmbH <oss@ew.tq-group.com>,
+ * D-82229 Seefeld, Germany.
+ * Author: Isaac L. L. Yuki
+ */
+//******************************************************************************
+
+#ifndef _APP_H_
+#define _APP_H_
+
+/*******************************************************************************
+ * Prototypes
+ ******************************************************************************/
+
+extern void BOARD_InitHardware(void);
+
+#endif /* _APP_H_ */