+++ CHANGELOG.md
@@ -14,6 +14,9 @@ files in git the submodule for meta-tq (meta layers meta-tq and meta-dumpling).
 
 ### Added
 
+* Initial board support for TQMa8MPxS on MB-SMARC-2
+* Initial board support for TQMa8MPxL on MBa8MPxL
+
 ### Changed
 
 * TQMa93xx: add and use board common pin_mux.c